requirements.txt
asgiref==3.7.2
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
cloudscraper==1.2.71
colorama==0.3.9
deckar01-ratelimit==3.0.2
deep-translator==1.11.4
deepl==1.15.0
Django==3.1.14
djangorestframework==3.14.0
idna==3.4
joblib==1.3.1
lxml==4.9.3
mysqlclient==2.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
wordhoard==1.4.9
